24 update

+++ app/Http/Controllers/ContratController.php
@@ -16,18 +16,27 @@ class ContratController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $contrats = Contrat::with(['orders','detail'])->get();
-        // dd($contrats);
-        return view('contrats.index',compact(['contrats']));
+    {   
+        if (getOnlineUser()->can('view-contrats')) 
+        {
+            $contrats = Contrat::with(['orders','detail'])->get();
+            // dd($contrats);
+            return view('contrats.index',compact(['contrats']));
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 
     public function details_index($id){
 
-        $contrat = Contrat::find($id);
-        $test_caterories = CategoryTest::all();
-        $details = Details_Contrat::where('contrat_id',$contrat->id)->get(); 
-        return view('contrats_details.index',compact(['contrat','details','test_caterories']));
+        if (getOnlineUser()->can('view-contrats')) {
+            $contrat = Contrat::find($id);
+            $test_caterories = CategoryTest::all();
+            $details = Details_Contrat::where('contrat_id',$contrat->id)->get(); 
+            return view('contrats_details.index',compact(['contrat','details','test_caterories']));
+        }
+       
+        return back()->with('error', "Vous n'êtes pas autorisé");
 
     }
 
@@ -80,30 +89,33 @@ class ContratController extends Controller
 
     public function details_store(Request $request){
 
-        $data = $this->validate($request, [
-            'contrat_id' => 'required|exists:contrats,id',
-            'pourcentage' => 'required',
-            'category_test_id' => 'required',
-        ]);
-
-        $contrat = Contrat::findorfail($data['contrat_id']);
-        $category_exit = $contrat->detail()->whereCategoryTestId($data['category_test_id'])->exists();
-
-        if ($category_exit) {
-            return back()->with('error', "Cette categorie existe ! ");
+        if ($request->user()->can('create-contrats')) {
+            $data = $this->validate($request, [
+                'contrat_id' => 'required|exists:contrats,id',
+                'pourcentage' => 'required',
+                'category_test_id' => 'required',
+            ]);
+    
+            $contrat = Contrat::findorfail($data['contrat_id']);
+            $category_exit = $contrat->detail()->whereCategoryTestId($data['category_test_id'])->exists();
+    
+            if ($category_exit) {
+                return back()->with('error', "Cette categorie existe ! ");
+            }
+    
+            try {
+    
+                DB::transaction(function () use ($data) {
+                    Details_Contrat::create($data);
+                });
+    
+                return back()->with('success', "Element enregistré avec succès ! ");
+    
+            } catch(\Throwable $ex){
+              return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+            }
         }
-
-        try {
-
-            DB::transaction(function () use ($data) {
-                Details_Contrat::create($data);
-            });
-
-            return back()->with('success', "Element enregistré avec succès ! ");
-
-        } catch(\Throwable $ex){
-          return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
-        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
 
     }
 
@@ -146,32 +158,32 @@ class ContratController extends Controller
      */
     public function update(Request $request)
     {
-
-        $data = $this->validate($request, [
-            'name2' => 'required',
-            'type2' => 'required',
-            'status2' => 'required',
-            'id2' => 'required',
-            'description2' => 'required',
-        ]);
-
-        try {
-            DB::transaction(function () use ($data) {
-
-                $contrat = Contrat::find($data['id2']);
-                $contrat->name = $data['name2'];
-                $contrat->type = $data['type2'];
-                $contrat->status = $data['status2'];
-                $contrat->description = $data['description2'];
-                $contrat->save();
-            });
-
-             return back()->with('success', "Mise à jour effectuée avec succès ! ");
-            } catch(\Throwable $ex){
-          return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
-      }
-
-
+        if ($request->user()->can('edit-contrats')) {
+            $data = $this->validate($request, [
+                'name2' => 'required',
+                'type2' => 'required',
+                'status2' => 'required',
+                'id2' => 'required',
+                'description2' => 'required',
+            ]);
+    
+            try {
+                DB::transaction(function () use ($data) {
+    
+                    $contrat = Contrat::find($data['id2']);
+                    $contrat->name = $data['name2'];
+                    $contrat->type = $data['type2'];
+                    $contrat->status = $data['status2'];
+                    $contrat->description = $data['description2'];
+                    $contrat->save();
+                });
+    
+                 return back()->with('success', "Mise à jour effectuée avec succès ! ");
+                } catch(\Throwable $ex){
+              return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+          }
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
 
     }
 
@@ -179,28 +191,29 @@ class ContratController extends Controller
     public function contrat_details_update(Request $request){
 
 
-        $data=$this->validate($request, [
-            'category_test_id2' => 'required',
-            'pourcentage2' => 'required',
-            'contrat_id2' => 'required',
-            'contrat_details_id2' => 'required',
-        ]);
-
-
-
-        try {
-            DB::transaction(function () use ($data) {
-                $contrat_detail = Details_Contrat::find($data['contrat_details_id2']);
-                $contrat_detail->category_test_id = $data['category_test_id2'];
-                $contrat_detail->pourcentage = $data['pourcentage2'];
-                $contrat_detail->save();
-            });
-
-             return back()->with('success', "Mise à jour effectué avec succès ! ");
-            } catch(\Throwable $ex){
-          return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
-      }
-
+        if (getOnlineUser()->can('edit-contrats')) {
+            $data=$this->validate($request, [
+                'category_test_id2' => 'required',
+                'pourcentage2' => 'required',
+                'contrat_id2' => 'required',
+                'contrat_details_id2' => 'required',
+            ]);
+    
+    
+    
+            try {
+                DB::transaction(function () use ($data) {
+                    $contrat_detail = Details_Contrat::find($data['contrat_details_id2']);
+                    $contrat_detail->category_test_id = $data['category_test_id2'];
+                    $contrat_detail->pourcentage = $data['pourcentage2'];
+                    $contrat_detail->save();
+                });
+    
+                 return back()->with('success', "Mise à jour effectué avec succès ! ");
+                } catch(\Throwable $ex){
+              return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+          }
+        }
 
     }
 
@@ -212,28 +225,41 @@ class ContratController extends Controller
      */
     public function destroy($id)
     {
-        $contrat = Contrat::find($id)->delete();
-        if ($contrat) {
-            return back()->with('success', "    Un élement a été supprimé ! ");
-        } else {
-            return back()->with('error', "    Ce contrat est utilisé ailleurs ! ");
+
+        if (getOnlineUser()->can('delete-contrats')) {
+            $contrat = Contrat::find($id)->delete();
+            if ($contrat) {
+                return back()->with('success', "    Un élement a été supprimé ! ");
+            } else {
+                return back()->with('error', "    Ce contrat est utilisé ailleurs ! ");
+            }
         }
         
+        return back()->with('error', "Vous n'êtes pas autorisé");
+  
     }
 
 
     public function destroy_details($id)
     {
-        Details_Contrat::find($id)->delete();
-        return back()->with('success', "    Un élement a été supprimé ! ");
+        if (getOnlineUser()->can('delete-contrats')) {
+            Details_Contrat::find($id)->delete();
+            return back()->with('success', "    Un élement a été supprimé ! ");
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 
     public function update_detail_status($id)
     {
-        $contrat = Contrat::findorfail($id);
-        $contrat->fill(["status" => "ACTIF"])->save();
+        if (getOnlineUser()->can('edit-contrats')) {
+            $contrat = Contrat::findorfail($id);
+            $contrat->fill(["status" => "ACTIF"])->save();
+    
+            return redirect()->route('contrats.index')->with('success', "Statut mis à jour ! ");
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
 
-        return redirect()->route('contrats.index')->with('success', "Statut mis à jour ! ");
     }
 
 }
