+++ app/Http/Controllers/HospitalController.php
@@ -14,8 +14,12 @@ class HospitalController extends Controller
      */
     public function index()
     {
-        $hopitals = Hospital::orderBy('name','asc')->get();
-        return view('hopitals.index',compact(['hopitals']));
+        if (getOnlineUser()->can('view-hopitaux')) {
+            $hopitals = Hospital::orderBy('name','asc')->get();
+            return view('hopitals.index',compact(['hopitals']));
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 
     /**
@@ -36,24 +40,26 @@ class HospitalController extends Controller
      */
     public function store(Request $request)
     {
-        $data=$this->validate($request, [
-            'name' => 'required',
-            'adresse' => 'nullable',    
-            'email' => 'nullable',        
-            'telephone' => 'required',  
-            'commission' => 'required |numeric|min:0|max:100',  
-        ]);
+        if (getOnlineUser()->can('create-hopitaux')) {
+            $data=$this->validate($request, [
+                'name' => 'required',
+                'adresse' => 'nullable',    
+                'email' => 'nullable',        
+                'telephone' => 'required',  
+                'commission' => 'required |numeric|min:0|max:100',  
+            ]);
 
-        
+            
 
-        try {
-            Hospital::create($data);
-            return back()->with('success', "Un hôpital a été enregistré ! ");
+            try {
+                Hospital::create($data);
+                return back()->with('success', "Un hôpital a été enregistré ! ");
 
-        } catch(\Throwable $ex){
-            return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+            } catch(\Throwable $ex){
+                return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+            }
         }
-
+        return back()->with('error', "Vous n'êtes pas autorisé");
     }
 
     /**
@@ -88,30 +94,34 @@ class HospitalController extends Controller
      */
     public function update(Request $request)
     {
-        $data=$this->validate($request, [
-            'name2' => 'required',
-            'id2' => 'required',
-            'adresse2' => 'nullable',    
-            'email2' => 'nullable',        
-            'telephone2' => 'required',  
-            'commission2' => 'required |numeric',  
-        ]);
+        if (getOnlineUser()->can('edit-hopitaux')) {
+            $data=$this->validate($request, [
+                'name2' => 'required',
+                'id2' => 'required',
+                'adresse2' => 'nullable',    
+                'email2' => 'nullable',        
+                'telephone2' => 'required',  
+                'commission2' => 'required |numeric',  
+            ]);
 
-        
-        try {
-            $hopital = Hospital::find($data['id2']);
-            $hopital->name = $data['name2'];
-            $hopital->adresse = $data['adresse2'];
-            $hopital->email = $data['email2'];
-            $hopital->telephone = $data['telephone2'];
-            $hopital->commission = $data['commission2'];
-            $hopital->save();
             
-            return back()->with('success', "Un hôpital a été mis à jour ! ");
+            try {
+                $hopital = Hospital::find($data['id2']);
+                $hopital->name = $data['name2'];
+                $hopital->adresse = $data['adresse2'];
+                $hopital->email = $data['email2'];
+                $hopital->telephone = $data['telephone2'];
+                $hopital->commission = $data['commission2'];
+                $hopital->save();
+                
+                return back()->with('success', "Un hôpital a été mis à jour ! ");
 
-        } catch(\Throwable $ex){
-            return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+            } catch(\Throwable $ex){
+                return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+            }
         }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
 
     }
 
@@ -123,7 +133,11 @@ class HospitalController extends Controller
      */
     public function destroy($id)
     {
-        Hospital::find($id)->delete();
-        return back()->with('success', "    Un élement a été supprimé ! ");
+        if (getOnlineUser()->can('delete-hopitaux')) {
+            Hospital::find($id)->delete();
+            return back()->with('success', "    Un élement a été supprimé ! ");
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 }
