+++ app/Http/Controllers/TestController.php
@@ -15,12 +15,16 @@ class TestController extends Controller
      */
     public function index()
     {
-        $tests = Test::all();
+        if (getOnlineUser()->can('view-examens')) {
+            $tests = Test::all();
    
-        $categories = CategoryTest::all();
+            $categories = CategoryTest::all();
+    
+            
+            return view('tests.index',compact(['tests','categories']));
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
 
-        
-        return view('tests.index',compact(['tests','categories']));
     }
 
     /**
@@ -41,7 +45,7 @@ class TestController extends Controller
      */
     public function store(Request $request)
     {
-      
+      if (getOnlineUser()->can('create-examens')) {
         $data = $this->validate($request, [
             'price' => 'required |numeric|gt:0',
             'name' => 'required |unique:tests,name',  
@@ -55,6 +59,9 @@ class TestController extends Controller
         } catch(\Throwable $ex){
             return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
         }
+      }
+      return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 
     /**
@@ -89,24 +96,28 @@ class TestController extends Controller
      */
     public function update(Request $request)
     {
-        $data=$this->validate($request, [
-            'id2' => 'required ', 
-            'price2' => 'required |numeric|gt:0',
-            'name2' => 'required ',  
-            'category_test_id2'=>'required'        
-        ]);
-
-        try {
-            $test = Test::find($data['id2']);
-            $test->name = $data['name2'];
-            $test->price = $data['price2'];
-            $test->category_test_id = $data['category_test_id2'];
-            $test->save();
-            return back()->with('success', " Mise à jour effectuée avec succès  ! ");
-
-        } catch(\Throwable $ex){
-            return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+        if (getOnlineUser()->can('edit-examens')) {
+            $data=$this->validate($request, [
+                'id2' => 'required ', 
+                'price2' => 'required |numeric|gt:0',
+                'name2' => 'required ',  
+                'category_test_id2'=>'required'        
+            ]);
+    
+            try {
+                $test = Test::find($data['id2']);
+                $test->name = $data['name2'];
+                $test->price = $data['price2'];
+                $test->category_test_id = $data['category_test_id2'];
+                $test->save();
+                return back()->with('success', " Mise à jour effectuée avec succès  ! ");
+    
+            } catch(\Throwable $ex){
+                return back()->with('error', "Échec de l'enregistrement ! " .$ex->getMessage());
+            }
         }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 
     /**
@@ -117,8 +128,12 @@ class TestController extends Controller
      */
     public function destroy($id)
     {
-       $test = Test::find($id)->delete();
+        if (getOnlineUser()->can('delete-examens')) {
+            $test = Test::find($id)->delete();
 
-        return back()->with('success', " Elément supprimé avec succès  ! ");
+            return back()->with('success', " Elément supprimé avec succès  ! ");
+        }
+        return back()->with('error', "Vous n'êtes pas autorisé");
+
     }
 }
